:sparkles: Refactor: policy class 이동
