+++ core/navigation/src/main/java/com/poptato/navigation/NavGraph.kt
@@ -10,7 +10,7 @@ import com.poptato.login.KaKaoLoginScreen
 import com.poptato.mypage.MyPageScreen
 import com.poptato.mypage.viewer.NoticeViewerScreen
 import com.poptato.mypage.viewer.FAQViewerScreen
-import com.poptato.mypage.viewer.PolicyViewerScreen
+import com.poptato.mypage.policy.PolicyViewerScreen
 import com.poptato.setting.servicedelete.ServiceDeleteScreen
 import com.poptato.setting.userdata.UserDataScreen
 import com.poptato.splash.SplashScreen
